Add LICENSE.md file
This license file contains the same content as preamble found in the
source files, however it also includes the following part which is
missing from the end (after "... CONTRACT, STRICT LIABILITY, OR TORT"):

> (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
> OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

I'm somewhat of two minds if it is okay to include this additional part,
however I think it is given that the package uses "BSD" for the license
value in setuptools config and the
`License :: OSI Approved :: BSD License` classifier and most BSD licenses
end with same text after TORT, and this includes all BSD licenses on the
OSI site.

The license text itself is however not the same as the the 3-Clause BSD license
found at any of these locations:

- https://opensource.org/licenses/BSD-3-Clause
- https://spdx.org/licenses/BSD-3-Clause.html
- https://directory.fsf.org/wiki/License:BSD-3-Clause
  (differs from first two anyway)

The difference is mainly the first part of the third clause, which in the
isodate license is "Neither the name of the authors nor the names of its
contributors" instead of "Neither the name of the copyright holder nor the
names of its contributors may" (SDPX, OSI) and "The name of the author"
(FSF).

The most popular project with similar license text that I could find is
https://github.com/andialbrecht/sqlparse.

I formatted the file with:

```sh
npx prettier --prose-wrap always LICENSE.md
```

Changes to `setup.cfg` come from `setup-cfg-fmt`.